+++ src/register/funnelPages/0_index.ts
@@ -1,13 +1,13 @@
-import 약관동의 from './약관동의';
-import 전화번호입력 from './전화번호입력';
-import 전화번호인증 from './전화번호인증';
-import 회사이메일입력 from './회사이메일입력';
-import 회사이메일인증 from './회사이메일인증';
-import 닉네임입력 from './닉네임입력';
-import 성별생년월일 from './성별생년월일';
-import SNS계정 from './SNS계정';
-import 기본프로필입력1 from './기본프로필입력1';
-import 기본프로필입력2 from './기본프로필입력2';
+import 약관동의 from './RegisterMemberAgreeFunnel';
+import 전화번호입력 from './RegisterEnterPhoneNumFunnel';
+import 전화번호인증 from './RegisterAuthenticatePhoneNumFunnel';
+import 회사이메일입력 from './RegisterEnterEmailFunnel';
+import 회사이메일인증 from './RegisterAuthenticateEmailFunnel';
+import 닉네임입력 from './RegisterEnterNicknameFunnel';
+import 성별생년월일 from './RegisterEnterGenderBirthFunnel';
+import SNS계정 from './RegisterSNSFunnel';
+import 기본프로필입력1 from './RegisterProfileFunnel1';
+import 기본프로필입력2 from './RegisterProfileFunnel2';
 
 export {
   약관동의,
