+++ src/register/funnelPages/RegisterEnterNicknameFunnel.tsx
